The stream names leave the catalog, because the file crossed 500 lines by one
`make craft-static` is red on main, and it is the whole tree rather than a diff,
so it takes `ci` down on every open pull request:

    backend/internal/shared/kernel/events/catalog.go
      1: [MAJOR] large-file — file is 501 lines (> 500) — split it by concept

Split by the concept the file already had two of. `streams.go` is what the
streams are CALLED and which one an event rides: the prefix, the entity
segments, the extension grammar, and the three functions that build key sets
from them. It is complete without knowing a single event type — the ops surface
enumerates it, the purge unlinks it, a consumer group declares its streams from
it. `catalog.go` keeps what it is named for: the closed set of types, and the
lookups over it.

Declarations moved, nothing rewritten. `streamOverlay` says it is named once and
`uniquenessclaims.txt` pins that claim by path, so the register entry follows it
to the new file — the same claim, still held, one line edited rather than added,
because that register only shrinks.
